overwrite

+++ src/snapshot.rs
@@ -184,7 +184,7 @@ fn path_resolve(path: String) -> String {
     full_path
 }
 
-pub fn export(snapshot: Snapshot, path: String) {
+pub fn export(snapshot: Snapshot, path: String, overwrite: bool) {
     let full_path = path_resolve(path);
 
     let mut fh: Vec<FileMetadata> = vec![];
@@ -205,17 +205,20 @@ pub fn export(snapshot: Snapshot, path: String) {
 
     let serialized = serde_json::to_string(&serializable).unwrap();
     // println!("{:#?}", serialized);
-
-    if Path::new(&full_path).exists() {
-        let _ = fs::remove_file(&full_path).unwrap();
-    }
-
-    // println!("{}", full_path);
-
     let filename = full_path.split('/').last().unwrap();
     let path_only = full_path.replace(filename, "");
-
+    // println!("{}", full_path);
     // println!("{}", path_only);
+
+    if Path::new(&full_path).exists() && overwrite {
+        let _ = fs::remove_file(&full_path).unwrap();
+        write_to_file(path_only, full_path, serialized);
+    } else if !Path::new(&full_path).exists() {
+        write_to_file(path_only, full_path, serialized);
+    }
+}
+
+fn write_to_file(path_only: String, full_path: String, serialized: String) {
     if let Ok(_) = fs::create_dir_all(path_only) {
         if let Ok(mut file_handle) = File::create(full_path) {
             file_handle.write_all(serialized.as_bytes()).unwrap()
@@ -313,8 +316,8 @@ mod tests {
     #[test]
     fn export_snapshot() {
         let test_snap = Snapshot::new(Path::new("/etc"), HashType::BLAKE3, vec![]);
-        export(test_snap.clone(), "/home/foxx/RustroverProjects/Fasching/output2/out.snapshot".to_string());
-        export(test_snap.clone(), "./output/out.snapshot".to_string());
+        export(test_snap.clone(), "/home/foxx/RustroverProjects/Fasching/output2/out.snapshot".to_string(), true);
+        export(test_snap.clone(), "./output/out.snapshot".to_string(), true);
 
 
     }
